Melhoria no tratamento de erro na exclusão da turma

+++ ceogestao/src/main/java/br/com/ceoestudos/ceogestao/dao/TurmaDAO.java
@@ -53,6 +53,12 @@ public class TurmaDAO {
         if(turma.getSituacao().equals(SituacaoTurma.EM_ANDAMENTO)){
             throw new RuntimeException("Não é possível excluir uma turma em andamento");
         }
+        if(turma.getContas()!=null && turma.getContas().size() > 0){
+            throw new RuntimeException("Não é possível excluir uma turma pois há contas relacionadas");
+        }
+        if(turma.getTratamentos()!=null && turma.getTratamentos().size() > 0){
+            throw new RuntimeException("Não é possível excluir uma turma pois há tratamentos relacionados");
+        }
         //turma = em.merge(turma);
         em.remove(turma);
     }
